Added html template for Google Sync Report. - Which has  2 tables  for enrolment and incoming calls. - created a new file for reuse of handlebar configs. - updated the sendEmail.js function to render html template instead of raw text. - created path of the syncReport template in constant.js
