+++ lib/constants.js
@@ -194,7 +194,8 @@ module.exports = {
             // to: ['amar17@navgurukul.org'],
             to: ['r@navgurukul.org', 'varatha@navgurukul.org'],
             cc: [],
-            subject: "Chanakya & gSheet just got sycned. Here is the report :)"
+            subject: "Chanakya & gSheet just got sycned. Here is the report :)",
+            syncReportHTMLReport: `${__dirname}/helpers/templates/syncReport.html`
         }
     },
     aws: {
