+++ lib/helpers/pdfGenerator.js
@@ -1,59 +1,13 @@
 'use strict';
-let webpage = require("webpage");
-let fs = require("fs");
-let handleBar = require("handlebars");
 let phantom = require("phantom");
-let system = require("system");
+let fs = require("fs");
+let handleBar = require("./handleBarsConfig");
 const CONSTANTS = require('../constants');
 const { uploadToS3 } = require('../helpers');
 const FormData = require("form-data");
 const _ = require("underscore");
 
-/* Random Handlebar Stuff */
-handleBar.registerHelper({
-	eq: function(v1, v2) {
-		return v1 === v2;
-	},
-	ne: function(v1, v2) {
-		return v1 !== v2;
-	},
-	lt: function(v1, v2) {
-		return v1 < v2;
-	},
-	gt: function(v1, v2) {
-		return v1 > v2;
-	},
-	lte: function(v1, v2) {
-		return v1 <= v2;
-	},
-	gte: function(v1, v2) {
-		return v1 >= v2;
-	},
-	and: function() {
-		return Array.prototype.slice.call(arguments).every(Boolean);
-	},
-	or: function() {
-		return Array.prototype.slice.call(arguments, 0, -1).some(Boolean);
-	}
-});
 
-handleBar.registerHelper("inc", function(value) {
-	return ++value;
-});
-
-handleBar.registerHelper("string_lower", function(index) {
-	let string_lower = "abcdefghijklmnopqrstuvwxyz";
-	return string_lower[index];
-});
-
-handleBar.registerHelper("string_upper", function(index) {
-	let string_upper = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-	return string_upper[index];
-});
-
-handleBar.registerHelper("hasParaTag", function(string){
-	return string.indexOf("<p>") > -1
-})
 
 const readFileAsync = filePath => {
 	return new Promise((resolve, reject) => {
@@ -202,10 +156,12 @@ exports.createTestAndAnswerKeyPDF = async function(questions, partnerName, asses
 		// });
 
 		// return the urls
-        let obj = {
+
+		let obj = {
             questionPaperUrl: _.where(urls, {pdfType: 'questionPaper'})[0].url,
             answerKeyUrl: _.where(urls, {pdfType: 'answerKey'})[0].url
-        }
+		}
+
         return obj
     });
 }
