+++ lib/helpers/sendEmail.js
@@ -2,6 +2,10 @@ const AWS = require('aws-sdk');
 // Set the region
 const CONSTANTS = require('../constants');
 const _ = require('underscore');
+const handleBar = require("./handleBarsConfig");
+const fs = require('fs');
+const { promisify } = require("util");
+
 
 const SESEmail = (receiverEmails, text, CcEmails = [], subject = '', isHtmlText = false) => {
 	AWS.config.update({
@@ -51,47 +55,6 @@ const SESEmail = (receiverEmails, text, CcEmails = [], subject = '', isHtmlText 
 		});
 };
 
-const getEnrolmentKeysErrorsMessage = platform => {
-	let errorMessages = '';
-	_.each(platform.enrolmentKeys, row => {
-		errorMessages += `\n Key ${row.key.key} has ${row.errors.length} problems: \n`;
-		_.each(row.errors, errorMessage => {
-			errorMessages += `\t- ${errorMessage}\n`;
-		});
-	});
-
-	return errorMessages;
-};
-
-const getIncomingCallsErrorsMessage = platform => {
-	let errorMessages = '';
-	_.each(platform.incomingCalls, row => {
-		errorMessages += `\n Incoming Call Id ${row.id} has ${row.errors.length} problems: \n`;
-		_.each(row.errors, errorMessage => {
-			errorMessages += `\t- ${errorMessage}\n`;
-		});
-	});
-
-	return errorMessages;
-};
-
-const getPlatformSyncErrorMessage = syncErrors => {
-	const { platform } = syncErrors;
-	let syncErrorReport = '';
-	totalErrors = platform.enrolmentKeys.length + platform.incomingCalls.length;
-
-	// errors while syncing chanakya from gSheet
-	if (totalErrors === 0) {
-		syncErrorReport = 'No Error in syncing chanakya from google Sheet.';
-	} else {
-		syncErrorReport += `Total ${totalErrors} rows had problem while syncing chanakya.\n`;
-		syncErrorReport += getEnrolmentKeysErrorsMessage(platform);
-		syncErrorReport += getIncomingCallsErrorsMessage(platform);
-	}
-
-	return syncErrorReport;
-};
-
 /**
  * 
  * @param {
@@ -109,18 +72,15 @@ const getPlatformSyncErrorMessage = syncErrors => {
  * } reports 
  */
 const sendEmailReport = reports => {
-	let emailReport = '';
-	emailReport += `Total enrolmentKeys Added ${reports.keysAdded}.\n`;
-	emailReport += `Total incomingCalls Added ${reports.incomingCallAdded}.\n`;
-	const { syncErrors } = reports;
-	emailReport += getPlatformSyncErrorMessage(syncErrors);
 
+	const source = fs.readFileSync(CONSTANTS.gSheet.emailReport.syncReportHTMLReport, "utf-8")
+	const htmlString = handleBar.compile(source)(reports)
 	const receiverEmails = CONSTANTS.gSheet.emailReport.to;
 	const CcEmails = CONSTANTS.gSheet.emailReport.cc;
 	const subject = CONSTANTS.gSheet.emailReport.subject;
+	return SESEmail(receiverEmails, htmlString, CcEmails, subject, true);
+}
 
-	return SESEmail(receiverEmails, emailReport, CcEmails, subject);
-};
 
 module.exports = {
 	sendEmailReport: sendEmailReport,
