+++ lib/helpers/syncGSheet.js
@@ -272,7 +272,8 @@ module.exports = class ChanakyaGSheetSync {
             let  dbCall = incomingCalls[call.id]
             let syncError = {
                 id: dbCall.id,
-                errors: []
+                errors: [],
+                row: call.row
             }
 
             let studentDetails = {}
@@ -307,8 +308,9 @@ module.exports = class ChanakyaGSheetSync {
         _.each(existingKeys, async (key) => {
             let  dbKey = enrolmentKeys[key.key]
             let syncError = {
-                key: dbKey,
-                errors: []
+                key: dbKey.key,
+                errors: [],
+                row: key.row
             }
 
             let studentDetails = {}
@@ -322,16 +324,16 @@ module.exports = class ChanakyaGSheetSync {
             // Recording the Actual Stage Column
             if (transitionResponse.shouldRecord === true) {
                 studentDetails['stage'] = key.student.actualStage
-                await this.assessmentService.patchStudentDetails(dbKey, studentDetails);
             } else if (transitionResponse.errorMessage !== null) {
                 // recording errors to send through email
                 syncError['errors'].push(transitionResponse.errorMessage)
             }
-
+            
             // Only the keys which has any errors
             if (syncError.errors.length > 0) {
                 this.reports.syncErrors.platform.enrolmentKeys.push(syncError)
             }
+            await this.assessmentService.patchStudentDetails(dbKey, studentDetails);
         });
     }
