+++ lib/helpers/templates/syncReport.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
+		<title>Sync Report</title>
+	</head>
+	<body>
+		<div>
+			<h3>Total enrolmentKeys Added {{ keysAdded }}.</h3>
+			<h3>Total incomingCalls Added {{ incomingCallAdded }}.</h3>
+		</div>
+		{{#if (ne (sum syncErrors.platform.enrolmentKeys.length syncErrors.platform.incomingCalls.length) 0)}}
+		<div>
+			<h3>
+				Total {{sum syncErrors.platform.enrolmentKeys.length syncErrors.platform.incomingCalls.length}} rows had
+				problem while syncing chanakya.
+			</h3>
+			<br>
+			<br>
+			{{#if (ne syncErrors.platform.enrolmentKeys.length 0)}}
+				<h3>Enrolment Keys Sync Report</h3>
+				<table style="width:80%" border="1">
+					<tr>
+						<th>Enrolment Key</th>
+						<th>Error in Syncing</th>
+						<th>gSheet Row #</th>
+					</tr>
+					{{#each syncErrors.platform.enrolmentKeys}}
+					<tr>
+						<td align="center">{{ this.key }}</td>
+						<td align="center">
+							{{#each this.errors}}
+							- {{this}} <br />
+							{{/each}}
+						</td>
+						<td align="center">{{ this.row }}</td>
+					</tr>
+					{{/each}}
+				</table>
+
+			{{/if}}
+			{{#if (ne syncErrors.platform.incomingCalls.length 0)}}
+				<h3>Incoming Call Sync Report</h3>
+				<table border="1" style="width:80%"> 
+					<tr>
+						<th>Incoming Call Ids</th>
+						<th>Error in Syncing</th>
+						<th>gSheet Row #</th>
+					</tr>
+					{{#each syncErrors.platform.incomingCalls}}
+					<tr>
+						<td align="center">{{ this.id }}</td>
+						<td align="center">
+							{{#each this.errors}}
+							- {{this}} <br />
+							{{/each}}
+						</td>
+						<td align="center">{{ this.row }}</td>
+					</tr>
+					{{/each}}
+				</table>
+			{{/if}}
+		</div>
+
+		{{else}}
+		<h3>No Error in syncing chanakya from google Sheet.</h3>
+		{{/if}}
+	</body>
+</html>
