+++ lib/services/assessment.js
@@ -321,7 +321,6 @@ module.exports = class AssessmentService extends Schmervice.Service {
                 isWhatsapp: true
             });
         } else {
-            console.log(studentId)
             await Contact.query(txn).patch({
                 isWhatsapp: true
             }).where({ studentId: studentId, mobile: whatsappNum });
@@ -355,7 +354,6 @@ module.exports = class AssessmentService extends Schmervice.Service {
 
     async patchStudentDetails(key, details={}, txn=null) {
         const { Student } = this.server.models();
-        console.log(details)
         // update stage of student if specified
         if (details.stage) {
             await this.recordStageTranisiton(key.student, details.stage, txn)
